# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Documentation for Warehouse Inventory Management System",
        "title": "Warehouse Inventory API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/history-stok": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of stock history with pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "History Stok"
                ],
                "summary": "Get all stock history",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.HistoryStokResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/history-stok/{barang_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get stock history for a specific barang with pagination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "History Stok"
                ],
                "summary": "Get stock history by barang ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Barang ID",
                        "name": "barang_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.HistoryStokResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.SpecificErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/stok": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of all stock items",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stok"
                ],
                "summary": "Get all stock",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MstokResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/stok/{barang_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get stock details for a specific barang",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stok"
                ],
                "summary": "Get stock by barang ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Barang ID",
                        "name": "barang_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.MstokResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "middleware.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "middleware.SpecificErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "models.BarangSimpleResponse": {
            "type": "object",
            "properties": {
                "kode_barang": {
                    "type": "string"
                },
                "nama_barang": {
                    "type": "string"
                }
            }
        },
        "models.BarangStokResponse": {
            "type": "object",
            "properties": {
                "harga_jual": {
                    "type": "number"
                },
                "kode_barang": {
                    "type": "string"
                },
                "nama_barang": {
                    "type": "string"
                },
                "satuan": {
                    "type": "string"
                }
            }
        },
...
        "models.HistoryStokResponse": {
            "type": "object",
            "properties": {
                "barang": {
                    "$ref": "#/definitions/models.BarangSimpleResponse"
                },
                "barang_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "jenis_transaksi": {
                    "type": "string"
                },
                "jumlah": {
                    "type": "integer"
                },
                "keterangan": {
                    "type": "string"
                },
                "stok_sebelum": {
                    "type": "integer"
                },
                "stok_sesudah": {
                    "type": "integer"
                },
                "user": {
                    "$ref": "#/definitions/models.UserSimpleResponse"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.MstokResponse": {
            "type": "object",
            "properties": {
                "barang": {
                    "$ref": "#/definitions/models.BarangStokResponse"
                },
                "barang_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "stok_akhir": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.UserSimpleResponse": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}